+++ knox/settings.py
@@ -9,6 +9,7 @@ USER_SETTINGS = getattr(settings, 'REST_KNOX', None)
 DEFAULTS = {
     'SECURE_HASH_ALGORITHM': 'cryptography.hazmat.primitives.hashes.SHA512',
     'AUTH_TOKEN_CHARACTER_LENGTH': 64,
+    'ENABLE_REFRESH_ENDPOINT': False,
     'TOKEN_TTL': timedelta(hours=10),
     'USER_SERIALIZER': None,
     'TOKEN_LIMIT_PER_USER': None,